playwright oh u funny thing
handle playwright now importing more files than before
make some files platform specific so playwright does not find them
fix tests when deprecating accessibilityLabel
