+++ packages/app/features/deposit/screen.tsx
@@ -53,7 +53,7 @@ function DepositWeb3Link() {
         {...useLink({
           href: '/deposit/web3',
         })}
-        accessibilityLabel="Deposit with Web3 Wallet"
+        aria-label="Deposit with Web3 Wallet"
       />
       <DepositStackSubheader>Web3 Wallet</DepositStackSubheader>
     </DepositStackButton>
